% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BLCA_auc}
\alias{BLCA_auc}
\title{Bladder Urothelial Carcinoma}
\format{A named numeric vector.}
\usage{
BLCA_auc
}
\description{
A vector of AUC computed with \code{\link[PAMES]{compute_AUC}} from PAMES
package.
}
\keyword{datasets}
