% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BRCA_auc}
\alias{BRCA_auc}
\title{Breast invasive carcinoma}
\format{A named numeric vector.}
\usage{
BRCA_auc
}
\description{
A vector of AUC computed with \code{\link[PAMES]{compute_AUC}} from PAMES
package.
}
\keyword{datasets}
