% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BRCA_islands}
\alias{BRCA_islands}
\title{Breast invasive carcinoma}
\format{A list of two vector of indexes (\code{hyper}- and \code{hypo}-methylated).}
\usage{
BRCA_islands
}
\description{
A list of informative CpG islands to be used with \code{\link[PAMES]{compute_purity}} from PAMES
package. Use with data from Bisulphite Sequencing after processing with
\code{\link[PAMES]{map_to_islands}}.
}
\keyword{datasets}
