% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COAD_auc}
\alias{COAD_auc}
\title{Colon adenocarcinoma}
\format{A named numeric vector.}
\usage{
COAD_auc
}
\description{
A vector of AUC computed with \code{\link[PAMES]{compute_AUC}} from PAMES
package.
}
\keyword{datasets}
