% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COAD_islands}
\alias{COAD_islands}
\title{Colon adenocarcinoma}
\format{A list of two vector of indexes (`hyper`- and `hypo`-methylated).}
\usage{
COAD_islands
}
\description{
A list of informative CpG islands to be used with `compute_purity` from PAMES
package. Use with data from Bisulphite Sequencing after processing with
`reduce_bs_data`.
}
\keyword{datasets}
