% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ESCA_auc}
\alias{ESCA_auc}
\title{Esophageal carcinoma}
\format{A named numeric vector.}
\usage{
ESCA_auc
}
\description{
A vector of AUC computed with \code{\link[PAMES]{compute_AUC}} from PAMES
package.
}
\keyword{datasets}
