% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KIRC_auc}
\alias{KIRC_auc}
\title{Kidney renal clear cell carcinoma}
\format{A named numeric vector.}
\usage{
KIRC_auc
}
\description{
A vector of AUC computed with \code{\link[PAMES]{compute_AUC}} from PAMES
package.
}
\keyword{datasets}
