% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KIRP_auc}
\alias{KIRP_auc}
\title{Kidney renal papillary cell carcinoma}
\format{A named numeric vector.}
\usage{
KIRP_auc
}
\description{
A vector of AUC computed with \code{\link[PAMES]{compute_AUC}} from PAMES
package.
}
\keyword{datasets}
