% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KIRP_sites}
\alias{KIRP_sites}
\title{Kidney renal papillary cell carcinoma}
\format{A list of two vector of indexes (`hyper`- and `hypo`-methylated).}
\usage{
KIRP_sites
}
\description{
A list of informative CpG sites to be used with `compute_purity` from PAMES
package. Use with data from Illumina BeadChip
}
\keyword{datasets}
