% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LUAD_auc}
\alias{LUAD_auc}
\title{Lung adenocarcinoma}
\format{A named numeric vector.}
\usage{
LUAD_auc
}
\description{
A vector of AUC computed with \code{\link[PAMES]{compute_AUC}} from PAMES
package.
}
\keyword{datasets}
