% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LUSC_auc}
\alias{LUSC_auc}
\title{Lung squamous cell carcinoma}
\format{A named numeric vector.}
\usage{
LUSC_auc
}
\description{
A vector of AUC computed with \code{\link[PAMES]{compute_AUC}} from PAMES
package.
}
\keyword{datasets}
