% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PAAD_auc}
\alias{PAAD_auc}
\title{Pancreatic adenocarcinoma}
\format{A named numeric vector.}
\usage{
PAAD_auc
}
\description{
A vector of AUC computed with \code{\link[PAMES]{compute_AUC}} from PAMES
package.
}
\keyword{datasets}
