% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PAAD_sites}
\alias{PAAD_sites}
\title{Pancreatic adenocarcinoma}
\format{A list of two vector of indexes (`hyper`- and `hypo`-methylated).}
\usage{
PAAD_sites
}
\description{
A list of informative CpG sites to be used with `compute_purity` from PAMES
package. Use with data from Illumina BeadChip
`reduce_bs_data`.
}
\keyword{datasets}
