% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAMESdata.R
\docType{package}
\name{PAMESdata-package}
\alias{PAMESdata}
\alias{PAMESdata-package}
\title{Precomputed list of informative CpG sites and CpG islands}
\description{
Using data from TCGA we copmute the sets of inforamtive CpG sites for 14 cancer.
I set were computed using the function `select_informative_sites`
from PAMES package.
}
\seealso{
\url{https://github.com/romagnolid/PAMES}
}
\author{
\strong{Maintainer}: Dario Last \email{first.last@example.com}

}
