% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAMESdata.R
\docType{package}
\name{PAMESdata}
\alias{PAMESdata}
\alias{PAMESdata-package}
\title{Precomputed sets of informative CpG sites and CpG islands}
\description{
Exploiting pulic data from from \href{https://portal.gdc.cancer.gov/}{TCGA},
we computed sets of informative
CpG sites and CpG islands for 14 cancer types.
}
\details{
After gathering data from different samples (see <paper waiting for publication>
for details) into a single matrix, CpG sites sets were created using
\code{\link[PAMES]{select_informative_sites}}.
To create CpG islands sets, the same matrixes were first remapped to CpG
islands using \code{\link[PAMES]{map_to_islands}} and then processed with
\code{\link[PAMES]{select_informative_islands}}.
}
\seealso{
\url{https://github.com/cgplab/PAMES}
}
