% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PRAD_auc}
\alias{PRAD_auc}
\title{Prostate adenocarcinoma}
\format{A named numeric vector.}
\usage{
PRAD_auc
}
\description{
A vector of AUC computed with \code{\link[PAMES]{compute_AUC}} from PAMES
package.
}
\keyword{datasets}
