% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{THCA_auc}
\alias{THCA_auc}
\title{Thyroid carcinoma}
\format{A named numeric vector.}
\usage{
THCA_auc
}
\description{
A vector of AUC computed with \code{\link[PAMES]{compute_AUC}} from PAMES
package.
}
\keyword{datasets}
