% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCEC_auc}
\alias{UCEC_auc}
\title{Uterine Corpus Endometrial Carcinoma}
\format{A named numeric vector.}
\usage{
UCEC_auc
}
\description{
A vector of AUC computed with \code{\link[PAMES]{compute_AUC}} from PAMES
package.
}
\keyword{datasets}
