% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LIHC_islands}
\alias{LIHC_islands}
\title{Liver hepatocellular carcinoma}
\format{A list of two vector of indexes (\code{hyper}- and \code{hypo}-methylated).}
\usage{
LIHC_islands
}
\description{
A list of informative CpG islands to be used with \code{\link[PAMES]{compute_purity}} from PAMES
package. Use with data from Bisulphite Sequencing after processing with
\code{\link[PAMES]{reduce_to_regions}}.
}
\keyword{datasets}
