% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PAAD_sites}
\alias{PAAD_sites}
\title{Pancreatic adenocarcinoma}
\format{A list of two vector of indexes (\code{hyper}- and \code{hypo}-methylated).}
\usage{
PAAD_sites
}
\description{
A list of informative CpG sites to be used with \code{\link[PAMES]{compute_purity}} from PAMES
package. Use with data from Illumina BeadChip.
\code{\link[PAMES]{remap_bs_data}}.
}
\keyword{datasets}
