% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCEC_islands}
\alias{UCEC_islands}
\title{Uterine Corpus Endometrial Carcinoma}
\format{A list of two vector of indexes (\code{hyper}- and \code{hypo}-methylated).}
\usage{
UCEC_islands
}
\description{
A list of informative CpG islands to be used with \code{\link[PAMES]{compute_purity}} from PAMES
package. Use with data from Bisulphite Sequencing after processing with
\code{\link[PAMES]{map_to_islands}}.
}
\keyword{datasets}
